removed task monitor and enabled background data save flow
